refactor(nestjs-signed-url): remove contract layer
