--- packages/nestjs-signed-url/src/contract/index.ts
@@ -1,2 +1,0 @@
-export type * from './options.js'
-export * from './provider.js'
