+++ packages/nestjs-signed-url/src/options.ts
