+++ packages/nestjs-signed-url/src/provider.ts
